Changed typedefs to work with 64-bit compiler.

+++ include/renderware.h
@@ -26,11 +26,11 @@ namespace rw {
 //typedef unsigned int uint;
 typedef char int8;
 typedef short int16;
-typedef long int32;
+typedef int int32;
 typedef long long int64;
 typedef unsigned char uint8;
 typedef unsigned short uint16;
-typedef unsigned long uint32;
+typedef unsigned int uint32;
 typedef unsigned long long uint64;
 typedef float float32;
